Add redirects to GitHub content.
We want to define pages at kubeflow.org/version/stable that can be updated
to point to the latest stable release.

We update stable to point at 0.2.2.

Note: When downloading deploy.sh you still need to set the environment variable
KUBEFLOW_VERSION because we don't set the appropriate version in the
deploy script.

Fix #1156