Add --zone option to kfctl generate step
It's a critical option to set at generate step.
The default value is us-east1-b.
If it's not specified at the generate step.  generated config files have the default zone hardcoded in multiple places.
Manually changing it cluster*.yaml and doing the apply doesn't seem to work.
Even if it did, it's better that the users are aware of this option when they are creating the deployment.