Refactor getting started - streamlined hierarchy (#1051)
* Refactor getting started - streamlined hierarchy

A number of changes / benefits:
1) eliminated some duplicate information
2) added three main top level areas for installation:
- cloud
- workstation / server
- existing kubernetes
3) Adding workstation operating systems - there are some slight
   variances in experience. For instance, for MacOS, we can emphasize
   brew to install some of the assets. For Windows, we'll want to add
   WSL2, and potentially chocolatey.

* Updated PR #1051 with recommended changes

This commit holds a set of minor changes that were requested in #1051

* Move VM based local workstation to reference

Moving the VM instructions out of getting started directory tree
so that the navigation bar remains clean / less busy.

* Add deleted assets files, reword k8s _index

A previous commit deleted the resources/_gen/assets/*/search*.content
file. This commit adds it back in.

Additionally, fixed the grammar in k8s/_index.md