Fixes kubeflow/website#1316 (#1620)
* Show which version of Kubeflow works for each sample project (#1316)

This commit will fetch the commit history from sample project to
determine when was the last time that the project received an
Updated.

* Fixed typo

This PR accidentally changed "repo" to "Rep". I've fixed that, and at the same time decided to use the full term instead of the original abbreviation.

* Removed misplaced line of text

This PR mistakenly copy-pasted a line of text from the financial time series sample into the issue summarization sample. I've removed the line.

Co-authored-by: Sarah Maddox <sarahmaddox@users.noreply.github.com>