Adding new section on managed certificates failure
This is a proposed fix for issue #1568.

I have added a new entry describing the steps on how to troubleshoot the certificate failure provided in the description of [issue](https://github.com/kubeflow/website/issues/1568). 

I also made just a couple of fixes in the section. 

I have a concern about the additional comment by @jlewi to this issue. It sounds too vague for me and most likely needs more detailed investigation and communication with SMEs.