Mention NVIDIA Tesla K80 Accelerators for the default GCP configuration
The default GCP configuration needs NVIDIA Tesla K80 Accelerators and zones which provide this accelerator need to be selected. This info is not mentioned in the main Deploy via CLI webpage and is in the Customizing Kubeflow on GKE webpage. Even though a user does not wish to customize their deployments, they need to know this info.