Fix failing test
Header rules test not passing.

In addition, according to [Kubeflow Versioning Policies](https://www.kubeflow.org/docs/reference/version-policy/), the alpha, beta and stable status are for applications, not deployment methods.

For some reason MicroK8s was the only deployment mechanism with this status, which is unfair, given the current investment and support we provide for Kubeflow on MicroK8s.