Refine Microk8s and Multipass docs (#2019)
* Refine Microk8s and Multipass docs

* Added reviewer suggestions

* Update getting-started-multipass.md

* Fix based on review

* Updating at par with #2043 on master branch

#2043 updated the docs for microk8s but for v1.0

I'm pulling forward the good changes to have the best docs :rocket:

* Removing out of date banner

* Refactor MicroK8s Multipass for Kubeflow guide

* Quick refactoring of last commit

* Fix failing test

Header rules test not passing.

In addition, according to [Kubeflow Versioning Policies](https://www.kubeflow.org/docs/reference/version-policy/), the alpha, beta and stable status are for applications, not deployment methods.

For some reason MicroK8s was the only deployment mechanism with this status, which is unfair, given the current investment and support we provide for Kubeflow on MicroK8s.

* Actually fixing the issue

* Upper and lower case - fix

* "Accessing" instead of "Access"

Co-authored-by: 8bitmp3 <19637339+8bitmp3@users.noreply.github.com>