fix broken Slack link
I noticed the kubeflow.org footer link for Slack did not work when I
first clicked on it, ie:
	"This link is no longer active"

Now, I was not a member of the Slack workspace when I first clicked
it.  Having subsequently found a working link elsewhere and
successfully signed up for the Slack workspace for Kubeflow, that
link does actually try to redirect me to the workspace.

I'm not sure how these redirects are implemented, or if this patch
represents an appropriate fix, but...

Using the url to the one listed for an invite in the
content/en/docs/about/contributing.md did succeed for me, both before
(ie: takes me to a sign up page) and after (takes me to the workspace)
membership.

Signed-off-by: Tim Pepper <tpepper@vmware.com>