Remove reference to insecure password-hashing site
The site referenced in the previous revision, passwordhashing.com, is profoundly insecure. It seems to the display last 10 BCrypt requests, *from any user worldwide*, in plain text. Clearly this is not an appropriate page to direct users.

Should references to this site be forcibly removed from older revisions of this document?