rm RFMVasconcelos (#2981)
Dear community,

In the last 2 years, I have had the great pleasure to be part of the community building this amazing open-source tool. I am quite satisfied in feeling that I left a small dent in the overall experience for new users.

Today I see my path take a different course, which I am excited to invest my full energy into. 🚀

I was also very lucky to work on Canonical's distribution - which I leave in the great hands of @knkski  @DomFleischmann and more. 

If you want to stay in touch, please connect via LinkedIn.

I nominate @knkski to take on some of the responsibilities I held, in his own capacity, given his 3 years of dedication to Kubeflow.

With Gratitude,
Rui