AI Platform Pipelines work with private GKE clusters
Update documentation to mention that AI Platform Pipelines 
work with private GKE clusters.
Kubeflow has many components and currently the
full deployment does not work on private GKE clusters.
However, the managed version of the component "Pipelines"
on GCP named "AI Platform Pipelines" does work with
private GKE cluster. This change makes it clear in the
documentation, in hopes to save time for developers
looking for the answer.