AI Platform Pipelines work with private GKE clusters (#3077)
* AI Platform Pipelines work with private GKE clusters 

Update documentation to mention that AI Platform Pipelines 
work with private GKE clusters.
Kubeflow has many components and currently the
full deployment does not work on private GKE clusters.
However, the managed version of the component "Pipelines"
on GCP named "AI Platform Pipelines" does work with
private GKE cluster. This change makes it clear in the
documentation, in hopes to save time for developers
looking for the answer.

* Fix spelling mistake

Co-authored-by: Shannon Bradshaw <shannon.bradshaw@arrikto.com>

* Update documentation based on reviewer feedback

Co-authored-by: Shannon Bradshaw <shannon.bradshaw@arrikto.com>