model-registry: update the installation guide (#4042)
I found installing model registry from the instructions on the website
to be rather hard to do. The instructions don't actually have all the
steps needed, so it sent me to the README in the manifests repo. Which,
confusingly, sent me right back to the website. Neither document is
entirely complete so I had to coble together bits from both places.

I'm assuming these instructions are primarily useful to existing
Kubeflow users who want to try out model registry so I've rewritten the
"Installing on Kubeflow Platform" section with the idea to get those
users up and running quickly. I'm further assuming that these users will
want the UI (I mean, it's a good UI, who wouldn't?) so we can cut the
"if you want the UI verbage" and simplify some of the commands.

Signed-off-by: Paul Boyd <pboyd@redhat.com>